Moved model code into code folder
