--- model/framework/neural_npfp/environment.yml
@@ -1,23 +1,0 @@
-name: neural_npfp_env
-channels:
-  - rdkit
-  - pytorch
-  - defaults
-dependencies:
-  - python=3.7.9
-  - pandas=1.1.3
-  - numpy=1.19.2
-  - scipy=1.5.2
-  - matplotlib=3.3.2=0
-  - seaborn=0.11.0=py_0
-  - pytables=3.6.1
-  - rdkit=2020.09.1.0
-  - yaml=0.2.5
-  - pyyaml=5.3.1
-  - pip
-  - pip:
-    - pyyml
-    - scikit-learn==0.23.2
-    - tqdm==4.51.0
-    - prettytable==2.1.0
-
