--- model/framework/neural_npfp/settings/settings.yml
@@ -1,34 +1,0 @@
-path_to_save: "../data/trained_models/npl_nonorm_64/"
-data:
-  batch_size: 128
-  fp_file: "../data/precomputed_fingerprints.csv"
-  remove_val_file: "../data/to_keep_molecules.pkl" 
-  descriptor_file: "../data/descriptors.csv"
-
-aux_model:
-  layers: [2048,1024,1024,64,48]
-  dropout: 0.2
-  lr: 0.0005
-  epochs: 20
-  baseline: False
-  with_npl: True
-  norm: 0
-baseline_model:
-  layers: [2048,1024,1024,64,48]
-  dropout: 0.2
-  lr: 0.0005
-  epochs: 20
-  baseline: True 
-  with_npl: True
-  norm: 0
-ae_model:
-  layers: [2048,512,64,512,2048]
-  dropout: 0.2
-  lr: 0.0001
-  epochs: 300
-  weights: [99/100,1/100]
-  with_npl: True
-  norm: 0  
-gcn_model:
-  test: [None]
-note: ""
