+++ model/framework/output.csv
@@ -1,0 +1,11 @@
+nn_naturalproductscore
+-1.6812998,-0.2782412,-0.33751443,-0.14361909,-0.06270826,0.010976896,0.0912635,0.41954118,0.3109002,-0.32007736,-0.6397441,0.24171363,0.5079252,0.31120142,0.66661215,-0.2963485,-0.5853054,-0.01711914,-0.019408122,-0.41957083,0.19611754,0.06760527,-0.86616856,0.022921726,-0.67604035,-0.078432664,-0.0031915456,0.36642322,0.18834925,0.34397346,-0.3273483,-0.54887813,0.36089265,0.7589592,0.9860836,-0.7498491,0.2716242,0.21025401,-0.120938,-0.19898154,0.108591646,0.523966,0.40453714,0.40872929,-0.2705711,-0.20214851,-1.5461602,-0.63551956,-1.0988039,0.41889974,-0.9898081,-0.69068265,0.2398473,0.35855675,-0.22391927,0.81176716,-0.26457104,0.47723454,-0.29137892,-0.035824582,0.3110468,-0.7718321,-0.4284699,0.47065517,-0.73932725
+9.593252,-1.0515856,1.3455393,-0.8681039,0.7911669,-0.4573663,0.7333934,-0.123563915,-0.11140619,0.049109116,0.11022642,2.0544944,0.08195042,-0.0155454725,-0.3453731,0.4085276,-0.24575384,1.5220003,-0.11212501,-0.08088289,0.18072778,0.16039582,-0.0050299354,0.0020703971,0.7642755,-1.6374018,0.00030349195,-0.22564048,-0.794091,0.4457339,-0.7633494,0.50851977,-1.2585192,0.26760596,-0.09336811,-0.32594693,0.07949917,-0.19285297,0.068422295,0.14005403,-0.56543875,0.08512534,-0.100482635,0.38859463,1.4161148,-0.6015644,-1.2693729,-0.4518423,-0.25787657,0.383785,-0.25614268,-0.36352697,-0.17958182,-1.2929475,1.1962535,-0.31076074,0.32863992,-0.19759528,0.17760389,1.6506944,0.6225522,-0.82426006,-0.42517412,0.5481632,0.123544425
+-2.9441888,0.23945138,-0.63876843,0.3402846,-0.11283603,-0.52732676,0.27279854,-0.074268,1.20929,0.07884595,-0.845742,0.12865086,0.17330195,-0.12845105,0.5003411,-0.24687158,-0.92258215,0.03572236,0.96255016,-0.4993853,0.59390914,-0.515653,-0.7641723,-1.0240963,-0.80927193,0.25186214,0.38035992,0.4116914,-0.4542784,-0.17430118,-0.39409935,-0.35951132,0.6545228,0.9646052,0.14822976,-1.318633,0.118215695,-0.67790556,-0.5171714,0.44498837,-0.420335,0.596558,-0.7624599,-0.119166136,-0.38674033,-0.61965257,-0.22829829,-0.8444584,0.11962101,-0.31424198,-0.52042687,-0.121337295,0.40260684,0.59973264,-0.29914328,-0.17739959,0.20066655,-0.10753064,-0.42565268,-0.5338508,0.3626438,0.13790943,-0.17795593,1.3235832,-1.3877566
+10.314021,1.5299251,-0.20646879,-0.09957216,0.7929567,-0.4569294,0.48848474,-0.81023204,-0.19399334,-0.098551765,-0.6684978,1.1558673,0.70010334,-0.5310082,0.1386774,0.32149696,-0.5001755,1.8702598,0.4025789,-1.0103604,0.8350768,-0.57917726,-0.43566397,0.013949186,-0.0584469,-1.4971721,0.35713783,-0.4150228,-0.677905,0.5972586,-0.65770376,0.12759674,-1.280208,0.07262649,0.5559223,-1.144632,-0.08710462,-0.06195645,-0.060779795,-0.1481574,-0.26263717,0.3406817,0.052050427,-0.67161417,1.4031985,-0.47949415,-0.24607027,-0.9860057,-0.043336734,0.03820303,0.15190554,0.5062883,-0.4031219,-1.2219393,1.1537164,0.10203602,0.06564111,0.12305437,0.19263218,0.34313264,-1.6644384,-0.35362178,-0.41500306,0.38198775,-0.34934062
+7.423603,1.4392507,-0.35306492,0.1327099,0.44630802,0.6296892,0.7876806,-0.6369245,-0.12784584,-0.025058791,-0.81185436,0.1338546,0.6094643,-0.6073948,-0.60651815,0.32811534,-0.24619237,1.4631956,-0.58726144,-0.719605,0.5314412,-0.9571647,-0.58644414,-0.26952335,-0.29289585,-0.91261905,-0.31948495,0.100691855,-0.48348618,0.9029935,-0.846491,-0.62304705,-1.1574942,0.6064203,1.4502009,-0.21676235,0.3010407,0.70336473,0.10401617,0.42976505,-1.2957188,1.1401433,-0.83724815,-1.233351,1.1699046,-0.121357754,-0.395171,-1.2283325,0.65585613,0.06199631,-0.034387365,0.28297827,0.7460605,-0.9192682,0.546065,0.60091674,-0.78702927,-0.16593273,-0.68526,0.45561162,-1.3227476,0.09041144,-0.7995501,0.7760018,-0.48275235
+4.8371525,0.92423606,-0.40362498,-0.14339904,0.79033995,-0.17136362,-0.3524512,-1.0556049,-0.08677308,0.3048208,-0.9389528,-0.1439344,0.20305133,0.15398407,-0.040053934,0.7756729,-1.1191787,1.3701303,-0.30862826,-0.8622101,0.7880907,-0.35516346,-0.32150227,-0.34216276,0.20618048,-0.64869505,0.3104831,-0.92116857,-0.78950477,0.48042792,-0.20383611,0.13587514,-0.82959664,-0.07912925,0.18898836,-0.34824052,-0.0511875,-0.1673469,-0.08172509,-0.0026527643,-0.062829405,0.16092072,-0.19493458,-0.54892373,1.0084778,-0.57898504,0.29378542,-0.66498816,-0.044950917,0.21065278,-0.17994004,0.59192514,-0.77272326,-0.29707277,0.30187264,0.2880518,0.79309523,0.29499835,-0.22868091,-0.24207386,-1.1442252,-0.06647056,-0.043316707,0.2614191,-0.25629133
+8.102048,1.5361587,-0.5450969,0.102879,0.43305427,-0.26207018,0.16817614,-0.010044366,0.1738381,-0.06475495,-0.3784704,0.27459893,0.73463106,-0.02889812,-0.0039831996,-0.3374058,-0.7246764,1.2279947,0.54645395,-0.73744434,-0.5871209,-0.3888914,-0.611248,-0.57071275,-1.0693204,-0.9278589,-0.028290138,0.20408314,0.025806159,0.3841822,0.095860034,-1.1971617,-0.9393742,0.5782856,1.2895017,-0.6281527,-0.0623935,0.47146723,-0.7075008,0.59173375,-0.18629971,0.41261655,0.78608114,-0.08170012,1.129457,-0.23305286,-0.44300058,-0.7002972,-0.7322571,0.09359269,-1.1480205,-0.4221998,-0.28582013,-1.0154164,0.16323236,0.8982577,-0.88218427,1.4973764,0.975533,1.3790598,-1.3123461,-0.29254046,-0.16478673,1.2745752,-1.0448015
+9.159543,1.3646432,0.6866678,0.83727586,-0.49551907,0.6231209,0.06419251,-0.41515058,0.78876346,0.6267855,-0.879182,0.64100933,-0.55733097,-0.04773131,0.4149241,0.017093405,0.12054658,1.1390874,1.1433568,0.8187202,0.36852038,-0.43958026,-1.1384492,-0.15957367,0.22763976,-1.3990092,-0.073031336,-0.006544724,-0.95897293,-0.022892158,-0.948441,0.650208,-0.9625218,0.13727203,0.28940305,0.41729838,0.8902857,-0.16642761,0.9159764,1.6018314,-0.76886225,-1.2572607,-1.1277108,0.16084178,1.2683271,-0.3992228,-0.118873745,0.37791383,1.3616337,0.19045222,-0.040088147,0.13669738,0.0795528,-1.1243031,-0.060124896,-1.0259457,0.041246086,0.050117895,-0.9367955,1.4729414,-1.1168171,0.9372357,-1.0148233,0.25828037,-0.7316733
+0.5848729,0.2760759,-0.62082684,0.8567698,-0.6173742,-0.45144433,-0.17691375,1.1738079,1.4931684,0.7881423,-0.30233145,0.14170153,0.8499404,1.1092957,1.2784746,0.14441323,-0.7377361,0.15592654,0.4861867,-0.19829221,1.6071856,-0.2484225,0.043350607,-0.70248544,-0.9675003,0.091091946,-0.9548074,0.8990557,0.9878489,-0.7629277,0.88184917,0.18695796,0.36356008,-0.43282524,0.29420948,0.4075787,-0.998542,0.2957476,-0.80193925,-0.25805712,0.50337505,-0.4492662,1.1553637,1.8199619,0.105713814,0.611178,-1.0080048,0.5602858,-1.3722446,0.83094895,-0.8002995,-0.10246049,0.55266887,0.22821566,-0.24449944,-0.059876323,-0.5268172,1.3323869,-0.9323892,0.041471347,-0.29930627,1.2637429,-1.294218,-0.103033215,-1.313833
+-8.877804,-0.6557944,-0.13719389,0.26710165,-0.32703933,0.15866819,0.14679834,-0.103214696,0.43162125,0.55351365,-0.58031875,-0.3428657,-0.08316608,0.10957559,0.8462586,0.35135838,-1.3011969,-0.777465,1.4061148,0.20592888,0.63411534,-0.7214279,-0.71004367,-1.1970924,-0.5264067,0.9844768,0.21552485,0.5492569,-0.5536266,0.05109454,-1.0551845,-0.053322688,1.3944917,0.68982655,0.02592744,0.22063963,0.2718534,-0.048968494,0.30071542,0.8880586,-0.7107699,-0.49525684,-0.45404702,0.118283994,-1.2299235,-1.0261754,-0.9182292,-0.391747,0.3914695,0.0064922273,-0.27553147,-0.44829106,0.020792872,1.4079058,-0.91565204,-0.7177969,-0.33432332,0.18916264,-0.6192729,-1.2746649,1.0531049,-0.14381328,-0.52875006,0.58844924,-0.6850431
