--- model/framework/run,sh
@@ -1,1 +1,0 @@
-python {0}/neural_npfp/neural_npfp/get_fp.py {1} {2} -s smiles
