+++ model/framework/sample.csv
@@ -1,0 +1,11 @@
+smiles
+Nc1nc(NC2CC2)c2ncn([C@H]3C=C[C@@H](CO)C3)c2n1
+C[C@]12CC[C@H]3[C@@H](CC=C4C[C@@H](O)CC[C@@]43C)[C@@H]1CC=C2c1cccnc1
+CC(=O)Nc1nnc(S(N)(=O)=O)s1
+CC(=O)O
+CC(=O)N[C@@H](CS)C(=O)O
+CC(=O)Oc1ccccc1C(=O)O
+Nc1nc(=O)c2ncn(COCCO)c2[nH]1
+O=C(O[C@H]1C[N+]2(CCCOc3ccccc3)CCC1CC2)C(O)(c1cccs1)c1cccs1
+CN(C)C/C=C/C(=O)Nc1cc2c(Nc3ccc(F)c(Cl)c3)ncnc2cc1O[C@H]1CCOC1
+CCCSc1ccc2nc(NC(=O)OC)[nH]c2c1
